check cmd and UVP item

+++ IN528ATE_tool/IN528ATE_tool/test_parameter.cs
@@ -48,7 +48,7 @@ namespace IN528ATE_tool
         /* Current Limit */
         public static double cv_setting;
         public static double cv_step;
-        public static int cv_wait;
+        public static double cv_wait;
 
     }
